sketch.js: Add visualizer for amplitude
Since the amplitude level is a pretty small number (for some reason!),
in order to make the circle look decent, I had to upscale it.

The result looks good for audio that has large differences in volume
over a very short period (happens a lot with bass-y kicks and snares).