sketch.js: Support uploading audio files (WIP)
Since multer handles `multipart/form-data`, the sketch file needs to add
an HTML form to the page and manually handle communication with the
backend.

But, the form positioning is still junk and the user doesn't get any
visual feedback after pressing the submit button.  For example: "did I
upload the correct file?" and "Can I start the new audio now?".